moves components inside of layouts
